0.101 - Входящее сообщение - кнопка "Переслать"

+++ EDI/web/src/main/java/view/documents/MessageCreate.java
@@ -93,12 +93,21 @@ public class MessageCreate extends HttpServlet {
                     req.setAttribute("userWhomList", Arrays.stream(documentEdi.getWhomString().split(",")).map(s -> (  s.matches("[0-9]+") ? UserImpl.INSTANCE.getUserById(Long.valueOf(s)) : null)).toArray(User[]::new));
                     req.setAttribute("uploadedFiles", UploadedFileServiceImpl.INSTANCE.getListByDocument(documentEdi));
                 }else {
+                    // Reply, Forward
                     Long documentCopyId = (Long) CommonModule.getNumberFromRequest(req, "documentCopyId", Long.class);
                     if (Objects.nonNull(documentCopyId)) {
                         AbstractDocumentEdi documentCopyEdi = AbstractDocumentEdiImpl.INSTANCE.getById(documentCopyId);
-                        req.setAttribute("theme", "Ответ на "+documentCopyEdi.getTheme());
-                        req.setAttribute("textInfo", CommonModule.getReplyString(documentCopyEdi.getText()));
-                        req.setAttribute("userWhomList", Collections.singletonList(documentCopyEdi.getAuthor()));
+                        String operationType = req.getParameter("operationType");
+                        if (operationType.equals("reply")){
+                            req.setAttribute("theme", "Ответ на "+documentCopyEdi.getTheme());
+                            req.setAttribute("textInfo", CommonModule.getReplyString(documentCopyEdi.getText()));
+                            req.setAttribute("userWhomList", Collections.singletonList(documentCopyEdi.getAuthor()));
+                        }else if (operationType.equals("forward")){
+                            req.setAttribute("theme", documentCopyEdi.getTheme());
+                            req.setAttribute("textInfo", "Пересылаю "+CommonModule.getReferenceOnDocument(documentCopyEdi));
+                            req.setAttribute("uploadedFiles", UploadedFileServiceImpl.INSTANCE.getListByDocument(documentCopyEdi));
+                            sessionDataElement.setDocumentCopyEdi(documentCopyEdi);
+                        }
                     }
                 }
 
@@ -184,7 +193,8 @@ public class MessageCreate extends HttpServlet {
                                     whomString = req.getParameter("post_users[]");
                                     whomString = (whomString.length() > 255 ? whomString.substring(0,254): whomString);
 
-                                    documentEdi = createOrUpdateDocument(req, documentEdi, timeStamp, currentUser, theme, textInfo, fileList, whomString, "save");
+                                    documentEdi = createOrUpdateDocument(req, documentEdi, timeStamp, currentUser, theme, textInfo, fileList, whomString, "save", sessionDataElement.getDocumentCopyEdi());
+                                    sessionDataElement.setDocumentCopyEdi(null);
                                     if (Objects.isNull(executorTask)) {
                                         executorTask = new ExecutorTask(timeStamp, null, true, currentUser, documentEdi, null, "", null, null, new java.sql.Timestamp(finalDate.getTime()), null, false, false, true);
                                         ExecutorTaskImpl.INSTANCE.save(executorTask);
@@ -205,7 +215,7 @@ public class MessageCreate extends HttpServlet {
 
                                     // --- Create document Message, business_process' classes: BusinessProcess, BusinessProcessSequence, ExecutorTask ---
 
-                                    documentEdi = createOrUpdateDocument(req, documentEdi, timeStamp, currentUser, theme, textInfo, fileList, whomString, "send");
+                                    documentEdi = createOrUpdateDocument(req, documentEdi, timeStamp, currentUser, theme, textInfo, fileList, whomString, "send", sessionDataElement.getDocumentCopyEdi());
                                     CommonBusinessProcessServiceImpl.INSTANCE.createAndStartBusinessProcess(currentUser, (Message) documentEdi, executorTask, timeStamp, usersIdArray, null, null, processTypeCommon, null, new java.sql.Timestamp(finalDate.getTime()));
                                     sessionDataElement.setElementStatus(ElementStatus.CLOSE);
 
@@ -256,7 +266,7 @@ public class MessageCreate extends HttpServlet {
         }
     }
 
-    private AbstractDocumentEdi createOrUpdateDocument(HttpServletRequest req, AbstractDocumentEdi documentEdi, java.sql.Timestamp timeStamp, User currentUser, String theme, String textInfo, List<UploadedFile> fileList, String whomString, String operationType) {
+    private AbstractDocumentEdi createOrUpdateDocument(HttpServletRequest req, AbstractDocumentEdi documentEdi, java.sql.Timestamp timeStamp, User currentUser, String theme, String textInfo, List<UploadedFile> fileList, String whomString, String operationType, AbstractDocumentEdi documentCopyEdi) {
 
         if (Objects.isNull(documentEdi)) {
             documentEdi = new Message(timeStamp, false, null, false, currentUser, CommonModule.getCorrectStringForWeb(theme), (operationType.equals("save") ? textInfo : CommonModule.getCorrectStringForWeb(textInfo)), whomString);
@@ -268,12 +278,25 @@ public class MessageCreate extends HttpServlet {
             MessageImpl.INSTANCE.update((Message) documentEdi);
         }
 
-        // files from request should be the same as the files in database - if this condition doesn't work we should delete files in database
-        List<UploadedFile> uploadedFilesFromRequest = getUploadedFileListFromRequest(req, documentEdi);
-        List<UploadedFile> filesInDatabase = UploadedFileServiceImpl.INSTANCE.getListByDocument(documentEdi);
-        filesInDatabase.removeAll(uploadedFilesFromRequest);
-        for (UploadedFile uploadedFile : filesInDatabase)
-            UploadedFileImpl.INSTANCE.delete(uploadedFile);
+        if (Objects.nonNull(documentCopyEdi)) {
+            List<UploadedFile> uploadedFilesForCopyFromRequest = getUploadedFileListFromRequest(req, documentCopyEdi);
+            for (UploadedFile uploadedFileForCopy : uploadedFilesForCopyFromRequest){
+                UploadedFile uploadedFileNew = new UploadedFile();
+                uploadedFileNew.setDocument(documentEdi);
+                uploadedFileNew.setFileName(uploadedFileForCopy.getFileName());
+                uploadedFileNew.setFilePath(uploadedFileForCopy.getFilePath());
+                uploadedFileNew.setName(uploadedFileForCopy.getName());
+                UploadedFileImpl.INSTANCE.save(uploadedFileNew);
+            }
+
+        }else {
+            // files from request should be the same as the files in database - if this condition doesn't work we should delete files in database
+            List<UploadedFile> uploadedFilesFromRequest = getUploadedFileListFromRequest(req, documentEdi);
+            List<UploadedFile> filesInDatabase = UploadedFileServiceImpl.INSTANCE.getListByDocument(documentEdi);
+            filesInDatabase.removeAll(uploadedFilesFromRequest);
+            for (UploadedFile uploadedFile : filesInDatabase)
+                UploadedFileImpl.INSTANCE.delete(uploadedFile);
+        }
 
         // add files to this document
         for (UploadedFile uploadedFile : fileList) {
